+++ CHANGES.md
@@ -21,6 +21,8 @@
   * Fix list-element extraction code, which asserted too strong a constraint. See issue #421
     for details. Thanks to Joel Burget for reporting.
 
+  * New bounded list functions: breverse/bsort. Contributed by Joel Burget.
+
 ### Version 7.12, 2018-09-23
 
   * Modifications to make SBV compile with GHC 8.6.1. (SBV should