+++ CHANGES.md
@@ -21,7 +21,8 @@
   * Fix list-element extraction code, which asserted too strong a constraint. See issue #421
     for details. Thanks to Joel Burget for reporting.
 
-  * New bounded list functions: breverse/bsort. Contributed by Joel Burget.
+  * New bounded list functions: breverse, bsort, bfoldrM, bfoldlM, and bmapM. Contributed by
+    Joel Burget.
 
   * Add a new puzzle example: "Documentation/SBV/Examples/Puzzles/LadyAndTigers.hs"
 
@@ -58,6 +59,9 @@
     supported.
 
 ### Version 7.10, 2018-07-20
+  * [BACKWARDS COMPATIBILITY] '==' and '/=' now always throw an error instead of
+    only throwing an error for non-concrete values.
+    https://github.com/LeventErkok/sbv/issues/301
 
   * [BACKWARDS COMPATIBILITY] Array declarations are reworked to take
     an initial value. The call 'newArray' now accepts an optional default